requirements.txt
discord
google-generativeai
pillow
re